Refactored the Scraping domain to have one root: scraping.all_scraping.scraping.py
