+++ scraping/all_scraping/ScrapeRss/scrape_rss.py
@@ -40,7 +40,7 @@ import sys
 import os
 
 # Connect to db_connector from parent directory
-PARENT_DIR = ".."
+PARENT_DIR = "../.."
 CURRENT_DIR = os.path.dirname(
     os.path.realpath(os.path.join(os.getcwd(), os.path.expanduser(__file__)))
 )
@@ -64,7 +64,7 @@ RSS_STACK = {}
 CACHE = set()
 
 # Database configurations
-from DatabaseConnector.db_connector import DatabaseConnector
+from scraping.DatabaseConnector.db_connector import DatabaseConnector
 
 db_connector = DatabaseConnector(config_path="./db.json")
 db_connector.connect()
@@ -73,30 +73,29 @@ db_connector.connect()
 db_connector_prodv2 = DatabaseConnector(config_path="./db.prodv2.json")
 db_connector_prodv2.connect()
 
-
 # import all news sources
-from ScrapeRss.rss_sites import NEWS_SOURCES
+from ..ScrapeRss.rss_sites import NEWS_SOURCES
 
 # NewsParser
-from ScrapeRss.NewsParser import NewsParser
+from ..ScrapeRss.NewsParser import NewsParser
 
 # NewsContent
-from ScrapeRss.NewsContent import NewsContent
+from ..ScrapeRss.NewsContent import NewsContent
 
 # Constant values
-from ScrapeRss.globals import CACHE_FILE, OUTPUT_FILENAME
-from ScrapeRss.globals import HEADER, THREAD_LIMIT
-from ScrapeRss.globals import DATE_FORMAT, CORONA_KEYWORDS, SPECIAL_LANG
-from ScrapeRss.globals import SEED_QUEUE, EXTRACT_QUEUE
+from ..ScrapeRss.globals import CACHE_FILE, OUTPUT_FILENAME
+from ..ScrapeRss.globals import HEADER, THREAD_LIMIT
+from ..ScrapeRss.globals import DATE_FORMAT, CORONA_KEYWORDS, SPECIAL_LANG
+from ..ScrapeRss.globals import SEED_QUEUE, EXTRACT_QUEUE
 
 # Helper functions
-from ScrapeRss.helpers import (
+from ..ScrapeRss.helpers import (
     get_seed_page,
     get_title_from_article,
     get_published_at_value,
     get_author_value,
 )
-from ScrapeRss.helpers import (
+from ..ScrapeRss.helpers import (
     attempt_extract_from_meta_data,
     corona_keyword_exists_in_string,
     extract_article,
@@ -193,9 +192,9 @@ def extract_worker():
 
         # If keyword doesn't exists in article, skip
         if (
-            not corona_keyword_exists_in_string(rss_record["description"].lower())
-            and not corona_keyword_exists_in_string(rss_record["title"].lower())
-            and not corona_keyword_exists_in_string(keywords.lower())
+                not corona_keyword_exists_in_string(rss_record["description"].lower())
+                and not corona_keyword_exists_in_string(rss_record["title"].lower())
+                and not corona_keyword_exists_in_string(keywords.lower())
         ):
             EXTRACT_QUEUE.task_done()
             continue
@@ -311,7 +310,7 @@ def write_to_cache(url):
         fh.write(url + "\n")
 
 
-if __name__ == "__main__":
+def scrape_rss():
     # arguments
     args = parser()
 
@@ -403,3 +402,7 @@ if __name__ == "__main__":
         count += len(rss_records)
     logging.debug("Total feeds: {}".format(count))
     logging.debug("Done scraping.")
+
+
+if __name__ == "__main__":
+    scrape_rss()
